requirements.txt
Django==2.1.2
paho-mqtt==1.4.0
pytz==2018.5
